refactor: move metric plugins under the dedicated subpackage
